crm_be/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-filter==22.1
drf-yasg==1.21.5
django-cors-headers==3.14.0
django-rest-swagger==2.2.0
django-import-export==3.2.0

python-decouple==3.6
python-dateutil==2.7.3
psycopg2-binary==2.9.5

# Export excel
openpyxl==3.1.2
XlsxWriter==3.1.9

# PDF
reportlab==3.6.12
PyPDF2==3.0.1
Japanera==1.1.0

# webserver
gunicorn==20.1.0